+++ xpeapp_ios/XpeApp/XpeApp/Newsletter/NewsletterPageView.swift
@@ -6,8 +6,11 @@
 //
 
 import SwiftUI
+import CoreData
 
 struct NewsletterPageView: View {
+    @Environment(\.managedObjectContext) var moc
+    @FetchRequest(sortDescriptors: []) private var cdNewsletters: FetchedResults<CDNewsletter>
     @State private var newsletters: [Newsletter] = []
     @State private var newsletterError = false
 
@@ -35,17 +38,12 @@ struct NewsletterPageView: View {
             .navigationTitle("Newsletters")
         }
         .onAppear {
-            Task {
-                if let newNs = await Newsletter.fetchAll() {
-                    newsletters = newNs.sorted(by: { $0.date > $1.date })
-                } else {
-                    newsletterError = true
-                }
+            // Note: This check is done in order not to refetch if state was kept
+            if newsletters.isEmpty {
+                onAppear()
             }
         }
-        .alert("Could not fetch newsletters", isPresented: $newsletterError) {
-            
-        }
+        .alert("Could not fetch newsletters", isPresented: $newsletterError) {}
     }
     
     struct NewsletterDetailsView: View {
@@ -82,6 +80,45 @@ struct NewsletterPageView: View {
             .navigationBarTitleDisplayMode(.inline)
             .alert("Incorrect newsletter URL", isPresented: $incorrectURL) {
                 
+            }
+        }
+    }
+    
+    private func onAppear() {
+        newsletters = cdNewsletters.map { cdn in
+            return Newsletter(
+                id: cdn.id ?? "",
+                date:cdn.date ?? Date(),
+                pdfUrl: cdn.pdfUrl ?? "",
+                publicationDate: cdn.publicationDate ?? Date(),
+                summary: cdn.summary ?? "")
+        }.sorted(by: {$0.date > $1.date})
+        
+        
+        Task {
+            guard let newNs = await Newsletter.fetchAll() else { return } // Todo: add error logging
+            DispatchQueue.main.async {
+                newsletters = newNs.sorted(by: {$0.date > $1.date})
+            }
+            
+            // Note: We repopulate the cache because the data could have
+            // changed in existing and new newsletters since last time
+            for cdn in cdNewsletters {
+                moc.delete(cdn)
+            }
+            
+            for n in newNs {
+                let cdn = CDNewsletter(context: moc)
+                cdn.id = n.id
+                cdn.date = n.date
+                cdn.pdfUrl = n.pdfUrl
+                cdn.publicationDate = n.publicationDate
+                cdn.summary = n.summary
+                do {
+                    try moc.save()
+                } catch {
+                    print("Error trying to recreate newsletter cache: \(error)")
+                }
             }
         }
     }